Reconfigured Startup Events and Cacheing
* Due to a lack of cache, MapNotes and Quests weren't always updated.
Setting up a one second delay for normal functions and running a
pre-caching loop of sorts through the QuestLog as a buffer seems to have
helped pull the data into the game clients local cache more completely
before allowing normal CheckLog and MapNotes functions. The pre-caching
loop isn't triggered by any event until after all the first round of
cache tables are built. Normal startup is resumed and the
QUEST_LOG_UPDATE and QUEST_ITEM_UPDATE events are registered. The
pre-caching loop continues to fire and is triggered by QLU and is
heavily throttled to only allow it to run once every 0.1 seconds, which
is enough to only allow it to fire once even when a large number of
QLU’s get triggered back to back.
* Most of the other quest and note management functions had already been
streamlined but I was able to remove some of them for a cleaner
environment. Some of the work flows even appear to be faster and most of
all the flickering is gone. During testing, the only time I really ever
noticed it is when I had showObjectives set to false and only allowed
Map Icons to appear when actively tracked. When toggled this way, I had
to delay the DRAWNOTES event due to the way the events were firing. If
MapNotes are always on, then I very rarely see any flickering. This is
the default setting that most users are used to using.
* Per your request the Questie:Toggle() now turns off the
Questie:OnUpdate and the Questie:OnEvent functions. I’ve tried it out
and it works. No events fire and no errors are triggered.