Minor Tweaks - Detect Quest Items Function
*Fixed a couple timing issues while playing on Elysium tonight centered
mainly around when and how MapNotes are redrawn after getting or
finishing a quest.
*Created the Quest Item detection stack so now only quest items looted
will trigger a QuestLog update. It looks in the QuestieCachedQuests
table and parses the objectives then compares against how many you need
to complete the quest. Once you obtain the needed amount, any further
updates are skipped.
*Discovered a nasty little bug centered around EQL3’s Auto Finish Quest
feature. If your inventory is full and you attempt to turn in a quest
that requires you to choose a reward or a quest that gives you a reward,
the QuestLog frame goes nuts with updates and the player is spammed in a
loop. Created an inventory checker and if the available slots are zero,
it toggles EQL3’s Auto Finish Quest feature off after saving its user
state so it doesn’t switch it off permanently if the user chooses to
have it on all the time. The player is then greeted with a Quest Error
Sound and issued an error message letting them know their inventory is
full, declines to finish the quest and exits the Quest Frame. Once the
player empties his inventory and attempts to hand in the quest again,
the check falls back to an else statement putting their EQL3 setting
back to its previously saved state and finishes the quest.