Tweaked Quest Fix and CheckQuestLog
*Short term fix is to show both locations for "Vial of Blessed Water" -
they are the same item so no way around it. Player will just have to
know that on Step 6 of Crown of the Earth they'll have to go to the
Orcacle Glade not Temple of the Moon. Maybe reference the tooltip?
*I had a delay placed in the wrong location in the CheckQuestLog
function. Some zones appear to be more latent than others and the extra
time is needed for it to make it into the QuestLog / QuestCache before
updating the MapNotes.